[BACKEND] Added route authorization for half of the route files, where needed

+++ backend/services/matchService.js
@@ -1,4 +1,7 @@
 const matchDAO = require('../DAO/matchDAO');
+const { SuperError, ERR_FORBIDDEN, ERR_NOT_FOUND } = require('../errors');
+const employerService = require('./employerService');
+const placementService = require('./placementService');
 const studentService = require('./studentService');
 
 module.exports = {
@@ -17,11 +20,29 @@ module.exports = {
         return matchDAO.getMatchesByStudentId(studentId);
     },
     
-    deleteMatch: async (studentId, placementId) => {
-        return await matchDAO.deleteMatch(studentId, placementId);
+    deleteMatch: async (studentId, placementId, auth) => {
+        let student = await studentService.getStudent(studentId);
+        let employer = await employerService.getEmployerByPlacementId(placementId);
+        //( foo && !bar ) || ( !foo && bar )
+        if (  (auth.id !== student.userId && auth.id == employer.userId) 
+        || (  (auth.id == student.userId) && auth.id !== employer.userId)) {
+            throw new SuperError(ERR_FORBIDDEN, 'You are not authorized to delete this match');
+            return;
+        }
+        return await matchDAO.deleteMatch(studentId, placementId);  
     },
 
-    getMatchesByPlacementId: async (placementId) => {        
+    getMatchesByPlacementId: async (placementId, auth) => {        
+        let employer = await employerService.getEmployerByPlacementId(placementId);
+        if(employer === null) {
+            throw new  SuperError(ERR_NOT_FOUND, 'The placement or the employer were not found');
+            return;
+        }
+        if (auth.id !== employer.userId) {
+            throw new SuperError(ERR_FORBIDDEN, 'You are not authorized to delete this match');
+            return;
+        }
+
         let matches = await matchDAO.getMatchesByPlacementId(placementId);
         
         let niceMatches = [];
